Did requested changes for PR and added a bandaid fix for the game dir detection in OE - can now be specified manually

+++ spt/OrangeBox/scripts/parsed_script.hpp
@@ -43,7 +43,7 @@ namespace scripts
 	private:
 		std::string saveName;
 		std::string scriptName;
-		long long int afterFramesTick;
+		int afterFramesTick;
 		Savestate GetSaveStateInfo();
 		std::vector<Savestate> saveStates;
 	};
